Refactor Kanban board components and implement drag-and-drop functionality
- Modularized the Kanban board by creating new components: KanbanColumnContent, KanbanDragOverlay, and a custom hook useKanbanDragDrop.
- Removed unused imports and consolidated drag-and-drop logic to improve code organization and maintainability.
- Enhanced task management with improved drag-and-drop interactions, including archiving and stage transitions.
