+++ client/src/components/KanbanBoard.tsx
@@ -1,36 +1,14 @@
-/* eslint-disable @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-explicit-any, @typescript-eslint/no-misused-promises, @typescript-eslint/no-floating-promises, @typescript-eslint/no-confusing-void-expression, @typescript-eslint/prefer-nullish-coalescing, @typescript-eslint/return-await, @typescript-eslint/no-unnecessary-condition, @typescript-eslint/no-unused-vars, @typescript-eslint/no-empty-function, @typescript-eslint/no-redundant-type-constituents, @typescript-eslint/no-unnecessary-type-conversion, @typescript-eslint/no-unnecessary-boolean-literal-compare, @typescript-eslint/require-await, @typescript-eslint/no-unused-expressions, @typescript-eslint/no-non-null-assertion, @typescript-eslint/prefer-optional-chain -- R2 baseline: strict fixes deferred to follow-up tasks */
-import { useEffect, useMemo, useState } from 'react';
-import { Task } from '@shared/schema';
+/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/prefer-nullish-coalescing -- R2 baseline: strict fixes deferred to follow-up tasks */
+import { useMemo } from 'react';
+import type { Task } from '@shared/schema';
 import { useStages, useSubStages } from '@/hooks/use-stages';
-import {
-  DndContext,
-  DragOverlay,
-  closestCenter,
-  pointerWithin,
-  rectIntersection,
-  CollisionDetection,
-  MouseSensor,
-  TouchSensor,
-  useSensor,
-  useSensors,
-  DragStartEvent,
-  DragOverEvent,
-  DragEndEvent,
-  MeasuringStrategy,
-} from '@dnd-kit/core';
-import { SortableContext, verticalListSortingStrategy } from '@dnd-kit/sortable';
+import { useKanbanDragDrop } from '@/hooks/use-kanban-drag-drop';
+import { DndContext, MeasuringStrategy } from '@dnd-kit/core';
 import { TaskColumn } from './TaskColumn';
-import { TaskCard } from './TaskCard';
-import { TaskCardSummary } from './TaskCardSummary';
+import { KanbanColumnContent } from './KanbanColumnContent';
+import { KanbanDragOverlay } from './KanbanDragOverlay';
 import { ArchiveZone } from './ArchiveZone';
-import { DayPlanSubStage } from './DayPlanSubStage';
-import { useUpdateTask, useArchiveTask } from '@/hooks/use-tasks';
-import {
-  DEFAULT_STAGE_COLORS,
-  getStatusFromStageName,
-  isInProgressStageName,
-} from '@shared/constants';
-import { cn } from '@/lib/utils';
+import { DEFAULT_STAGE_COLORS } from '@shared/constants';
 
 export interface KanbanBoardProps {
   tasks: Task[];
@@ -45,233 +23,29 @@ export function KanbanBoard({
   viewMode = 'detail',
   focusMode = false,
 }: KanbanBoardProps) {
-  const updateTask = useUpdateTask();
-  const archiveTask = useArchiveTask();
-  const [activeId, setActiveId] = useState<number | null>(null);
-  const [activeTasks, setActiveTasks] = useState(tasks);
-  const [isOverArchive, setIsOverArchive] = useState(false);
-
   const { data: stages = [] } = useStages();
-
   const { data: allSubStages = [] } = useSubStages();
 
-  useEffect(() => {
-    setActiveTasks(tasks);
-  }, [tasks]);
-
-  // Custom collision detection: prioritize archive zone using both pointer-within
-  // and rect-intersection checks, then pointerWithin for columns, then closestCenter.
-  // Using pointerWithin first for archive ensures detection when the pointer is
-  // directly over the zone, regardless of scroll position or dragged element size.
-  const collisionDetection: CollisionDetection = (args) => {
-    const { droppableContainers } = args;
-    const archiveContainers = droppableContainers.filter(
-      (c) => c.id === 'archive' || String(c.id) === 'archive',
-    );
-    if (archiveContainers.length > 0) {
-      // Check pointer position first — most reliable for archive zone detection
-      // across scroll containers and regardless of dragged element size.
-      const archivePointerCollisions = pointerWithin({
-        ...args,
-        droppableContainers: archiveContainers,
-      });
-      if (archivePointerCollisions.length > 0) {
-        return archivePointerCollisions;
-      }
-      // Fall back to rect intersection for archive zone (element overlap)
-      const archiveRectCollisions = rectIntersection({
-        ...args,
-        droppableContainers: archiveContainers,
-      });
-      if (archiveRectCollisions.length > 0) {
-        return archiveRectCollisions;
-      }
-    }
-    const pointerCollisions = pointerWithin(args);
-    if (pointerCollisions.length > 0) {
-      return pointerCollisions;
-    }
-    return closestCenter(args);
-  };
-
-  // Separate sensors for mouse and touch to avoid conflicts
-  // TouchSensor uses delay so scrolling still works (long-press to drag)
-  // MouseSensor uses distance for desktop testing
-  const sensors = useSensors(
-    useSensor(TouchSensor, {
-      activationConstraint: { delay: 200, tolerance: 5 },
-    }),
-    useSensor(MouseSensor, {
-      activationConstraint: { distance: 8 },
-    }),
-  );
-
-  function handleDragStart(event: DragStartEvent) {
-    setActiveId(event.active.id as number);
-    // Haptic feedback on drag start
-    if ('vibrate' in navigator) navigator.vibrate(15);
-  }
-
-  function handleDragOver(event: DragOverEvent) {
-    const { over } = event;
-    const isArchive =
-      over?.id === 'archive' ||
-      String(over?.id) === 'archive' ||
-      over?.data?.current?.type === 'archive';
-    setIsOverArchive(!!isArchive);
-  }
-
-  function handleDragEnd(event: DragEndEvent) {
-    const { active, over } = event;
-    setIsOverArchive(false);
-
-    if (!over) {
-      setActiveId(null);
-      return;
-    }
-
-    const activeTask = activeTasks.find((t) => t.id === active.id);
-
-    const isArchive =
-      over.id === 'archive' ||
-      String(over.id) === 'archive' ||
-      over.data?.current?.type === 'archive';
-    if (isArchive && activeTask) {
-      if ('vibrate' in navigator) navigator.vibrate([10, 50, 10]);
-      setActiveTasks((tasks) => tasks.filter((t) => t.id !== activeTask.id));
-      archiveTask.mutate(activeTask.id, {
-        onError: () => {
-          setActiveTasks((tasks) => [...tasks, activeTask]);
-        },
-      });
-      setActiveId(null);
-      return;
-    }
-
-    const overContainerId = over.data.current?.sortable?.containerId || over.id;
-    const overId = typeof overContainerId === 'string' ? overContainerId : overContainerId;
-
-    const isSubStage =
-      over.data?.current?.type === 'SubStage' ||
-      (typeof overId === 'string' && /^\d+-\w+/.test(overId));
-
-    if (activeTask && isSubStage) {
-      let stageId: number;
-      let subStageTag: string;
-
-      if (over.data?.current?.type === 'SubStage') {
-        subStageTag = over.data.current.subStageTag;
-        const match = /^(\d+)-(.+)$/.exec(String(over.id));
-        stageId = match ? parseInt(match[1]!) : activeTask.stageId;
-      } else if (typeof overId === 'string') {
-        const match = /^(\d+)-(.+)$/.exec(overId);
-        if (match) {
-          stageId = parseInt(match[1]!);
-          subStageTag = match[2]!;
-        } else {
-          stageId = activeTask.stageId;
-          subStageTag = '';
-        }
-      } else {
-        stageId = activeTask.stageId;
-        subStageTag = '';
-      }
-
-      const stageSubStages = allSubStages.filter((ss: any) => ss.stageId === stageId);
-      const stageSubStageTags = stageSubStages.map((ss: any) => ss.tag);
-      const currentTags = activeTask.tags || [];
-      const filteredTags = currentTags.filter((tag: string) => !stageSubStageTags.includes(tag));
-      const newTags = [...filteredTags, subStageTag];
-
-      const targetStage = sortedStages.find((s: any) => s.id === stageId);
-      if (targetStage && activeTask.stageId !== stageId) {
-        const newStatus = getStatusFromStageName(targetStage.name);
-        setActiveTasks((tasks) =>
-          tasks.map((t) =>
-            t.id === activeTask.id ? { ...t, stageId, status: newStatus, tags: newTags } : t,
-          ),
-        );
-        updateTask.mutate({ id: activeTask.id, stageId, status: newStatus, tags: newTags });
-      } else {
-        setActiveTasks((tasks) =>
-          tasks.map((t) => (t.id === activeTask.id ? { ...t, tags: newTags } : t)),
-        );
-        updateTask.mutate({ id: activeTask.id, tags: newTags });
-      }
-
-      setActiveId(null);
-      return;
-    }
-
-    // overContainerId may be a number (from useDroppable) or string (from SortableContext)
-    const parsedStageId =
-      typeof overContainerId === 'number'
-        ? overContainerId
-        : typeof overContainerId === 'string' && /^\d+$/.test(overContainerId)
-          ? parseInt(overContainerId, 10)
-          : null;
-
-    if (activeTask && parsedStageId !== null) {
-      const newStageId = parsedStageId;
-      const newStage = sortedStages.find((s: any) => s.id === newStageId);
-
-      if (activeTask.stageId !== newStageId && newStage) {
-        const newStatus = getStatusFromStageName(newStage.name);
-        let newTags = activeTask.tags || [];
-        const newStageSubStages = allSubStages.filter((ss: any) => ss.stageId === newStageId);
-        if (newStageSubStages.length === 0) {
-          const allSubStageTags = allSubStages.map((ss: any) => ss.tag);
-          newTags = newTags.filter((tag: string) => !allSubStageTags.includes(tag));
-        } else {
-          const otherStageSubStageTags = allSubStages
-            .filter((ss: any) => ss.stageId !== newStageId)
-            .map((ss: any) => ss.tag);
-          const newStageSubStageTags = newStageSubStages.map((ss: any) => ss.tag);
-          newTags = newTags.filter(
-            (tag: string) =>
-              !otherStageSubStageTags.includes(tag) || newStageSubStageTags.includes(tag),
-          );
-        }
-
-        setActiveTasks((tasks) =>
-          tasks.map((t) =>
-            t.id === activeTask.id
-              ? {
-                  ...t,
-                  stageId: newStageId,
-                  status: newStatus,
-                  tags: newTags.length > 0 ? newTags : null,
-                }
-              : t,
-          ),
-        );
-        updateTask.mutate({
-          id: activeTask.id,
-          stageId: newStageId,
-          status: newStatus,
-          tags: newTags.length > 0 ? newTags : null,
-        });
-      }
-    }
-
-    setActiveId(null);
-  }
-
-  const getDefaultStageColor = (index: number): string =>
-    DEFAULT_STAGE_COLORS[index % DEFAULT_STAGE_COLORS.length]!;
-
-  const sortedStages = useMemo(
-    () => [...stages].sort((a: any, b: any) => a.order - b.order),
-    [stages],
-  );
+  const sortedStages = useMemo(() => [...stages].sort((a, b) => a.order - b.order), [stages]);
 
   const stageColorMap = useMemo(() => {
     const map = new Map<number, string>();
-    sortedStages.forEach((stage: any, index: number) =>
-      map.set(stage.id, stage.color || getDefaultStageColor(index)),
+    sortedStages.forEach((stage, index) =>
+      map.set(stage.id, stage.color || DEFAULT_STAGE_COLORS[index % DEFAULT_STAGE_COLORS.length]),
     );
     return map;
   }, [sortedStages]);
+
+  const {
+    activeId,
+    activeTasks,
+    isOverArchive,
+    sensors,
+    collisionDetection,
+    handleDragStart,
+    handleDragOver,
+    handleDragEnd,
+  } = useKanbanDragDrop({ tasks, sortedStages, allSubStages });
 
   return (
     <DndContext
@@ -286,119 +60,27 @@ export function KanbanBoard({
         },
       }}
     >
-      {/* Mobile: vertical stack of full-width columns */}
       <div className="flex flex-col h-full gap-3 pb-4">
-        {sortedStages.map((stage: any) => {
+        {sortedStages.map((stage) => {
           const stageColor = stageColorMap.get(stage.id) || DEFAULT_STAGE_COLORS[0];
+          const stageTasks = activeTasks.filter((t) => t.stageId === stage.id);
           return (
             <TaskColumn
               key={stage.id}
               id={stage.id}
               title={stage.name}
-              count={activeTasks.filter((t) => t.stageId === stage.id).length}
+              count={stageTasks.length}
               stageColor={stageColor}
             >
-              <SortableContext
-                id={String(stage.id)}
-                items={activeTasks.filter((t) => t.stageId === stage.id).map((t) => t.id)}
-                strategy={verticalListSortingStrategy}
-              >
-                {(() => {
-                  const stageTasks = activeTasks.filter((task) => task.stageId === stage.id);
-
-                  const stageSubStages = allSubStages
-                    .filter((ss: any) => ss.stageId === stage.id)
-                    .sort((a: any, b: any) => a.order - b.order)
-                    .map((ss: any) => ({
-                      name: ss.name,
-                      tag: ss.tag,
-                      bgClass: ss.bgClass,
-                      opacity: ss.opacity / 100,
-                    }));
-
-                  if (stageSubStages.length > 0) {
-                    const stageSubStageTags = stageSubStages.map((ss: { tag: string }) => ss.tag);
-                    const tasksWithMatchingTags: Task[] = [];
-                    const tasksWithoutMatchingTags: Task[] = [];
-
-                    stageTasks.forEach((task) => {
-                      const tags = task.tags || [];
-                      if (tags.some((tag: string) => stageSubStageTags.includes(tag)))
-                        tasksWithMatchingTags.push(task);
-                      else tasksWithoutMatchingTags.push(task);
-                    });
-
-                    return (
-                      <div className="flex flex-col gap-2 min-h-[60px]">
-                        {stageSubStages.map(
-                          (
-                            subStage: {
-                              name: string;
-                              tag: string;
-                              bgClass: string;
-                              opacity: number;
-                            },
-                            subIndex: number,
-                          ) => {
-                            const subStageTasks = tasksWithMatchingTags.filter((task) =>
-                              (task.tags || []).includes(subStage.tag),
-                            );
-                            const finalTasks =
-                              subIndex === 0
-                                ? [...subStageTasks, ...tasksWithoutMatchingTags]
-                                : subStageTasks;
-
-                            return (
-                              <DayPlanSubStage
-                                key={subStage.tag}
-                                stageId={stage.id}
-                                stageName={stage.name}
-                                subStage={subStage}
-                                tasks={finalTasks}
-                                stageColor={stageColor}
-                                viewMode={viewMode}
-                                onTaskClick={onTaskClick}
-                              />
-                            );
-                          },
-                        )}
-                      </div>
-                    );
-                  }
-
-                  const inProgress = isInProgressStageName(stage.name);
-                  return viewMode === 'detail' ? (
-                    <div className="flex flex-col gap-2 min-h-[60px]">
-                      {stageTasks.map((task) => (
-                        <TaskCard
-                          key={task.id}
-                          task={task}
-                          onClick={onTaskClick}
-                          stageColor={stageColor}
-                          onInlineEdit={() => {}}
-                        />
-                      ))}
-                    </div>
-                  ) : (
-                    <div
-                      className={cn(
-                        'min-h-[60px] gap-2',
-                        inProgress ? 'flex flex-col' : 'flex flex-wrap content-start',
-                      )}
-                    >
-                      {stageTasks.map((task) => (
-                        <TaskCardSummary
-                          key={task.id}
-                          task={task}
-                          onClick={onTaskClick}
-                          stageColor={stageColor}
-                          isInProgress={inProgress}
-                        />
-                      ))}
-                    </div>
-                  );
-                })()}
-              </SortableContext>
+              <KanbanColumnContent
+                stageId={stage.id}
+                stageName={stage.name}
+                stageTasks={stageTasks}
+                allSubStages={allSubStages}
+                stageColor={stageColor}
+                viewMode={viewMode}
+                onTaskClick={onTaskClick}
+              />
             </TaskColumn>
           );
         })}
@@ -406,32 +88,13 @@ export function KanbanBoard({
         <ArchiveZone isOver={isOverArchive} />
       </div>
 
-      <DragOverlay>
-        {activeId
-          ? (() => {
-              const activeTask = activeTasks.find((t) => t.id === activeId);
-              if (!activeTask) return null;
-              const activeStageColor =
-                stageColorMap.get(activeTask.stageId) || DEFAULT_STAGE_COLORS[0];
-              const activeStage = sortedStages.find((s: any) => s.id === activeTask.stageId);
-              const activeStageName = activeStage?.name ?? '';
-              return (
-                <div className="opacity-80 rotate-1 cursor-grabbing">
-                  {viewMode === 'detail' ? (
-                    <TaskCard task={activeTask} onClick={() => {}} stageColor={activeStageColor} />
-                  ) : (
-                    <TaskCardSummary
-                      task={activeTask}
-                      onClick={() => {}}
-                      stageColor={activeStageColor}
-                      isInProgress={isInProgressStageName(activeStageName)}
-                    />
-                  )}
-                </div>
-              );
-            })()
-          : null}
-      </DragOverlay>
+      <KanbanDragOverlay
+        activeId={activeId}
+        activeTasks={activeTasks}
+        stageColorMap={stageColorMap}
+        sortedStages={sortedStages}
+        viewMode={viewMode}
+      />
     </DndContext>
   );
 }
